+++ src/components/Button/BackButton.tsx
@@ -1,5 +1,6 @@
 import { Feather } from '@expo/vector-icons';
-import { Text, View, StyleSheet } from 'react-native';
+import { Text, View } from 'react-native';
+import { backButtonStyles as styles } from './style';
 
 export const BackButton = ({ onPress }: { onPress: () => void }) => {
   return (
@@ -11,13 +12,4 @@ export const BackButton = ({ onPress }: { onPress: () => void }) => {
     </View>
   );
 };
-const styles = StyleSheet.create({
-  backButton: {
-    flexDirection: 'row',
-    paddingLeft: 20,
-  },
-  backButtonText: {
-    color: '#007AFF',
-    marginLeft: 4,
-  },
-});
+
