+++ src/components/Button/Button.tsx
@@ -1,0 +1,16 @@
+import { forwardRef } from 'react';
+import { Text, TouchableOpacity, TouchableOpacityProps } from 'react-native';
+import { styles } from './style';
+
+type ButtonProps = {
+  onPress?: TouchableOpacityProps['onPress'];
+  title?: string;
+} & TouchableOpacityProps;
+
+export const Button = forwardRef<TouchableOpacity, ButtonProps>(({ onPress, title }, ref) => {
+  return (
+    <TouchableOpacity ref={ref} style={styles.button} onPress={onPress}>
+      <Text style={styles.buttonText}>{title}</Text>
+    </TouchableOpacity>
+  );
+});
