+++ src/components/Button/index.ts
@@ -1,0 +1,4 @@
+import { BackButton } from './BackButton';
+import { Button } from "./Button";
+
+export { Button, BackButton };
